half-implemented history on server

+++ hoive/src/game/board.rs
@@ -13,7 +13,7 @@ use crate::maths::coord::Spiral;
 #[derive(Debug, Eq, PartialEq, Clone)]
 pub struct Board<T: Coord> {
     pub chips: HashMap<Chip, Option<T>>, // player chips (both teams)
-    pub turns: u32,                      // number of turns that have elapsed
+    pub turns: usize,                      // number of turns that have elapsed
     pub coord: T,                        // coordinate sytem for the board e.g. Cube, HECS
     pub history: History,                // record of all previous moves
     pub size: i8,                        // the size of the board in dheight
